Refactor java client code
