+++ java-client/src/main/java/com/wavefront/integrations/WavefrontDirectSender.java
@@ -1,8 +1,5 @@
 package com.wavefront.integrations;
 
-import com.wavefront.api.DataIngesterAPI;
-import com.wavefront.common.NamedThreadFactory;
-
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,22 +7,13 @@ import org.slf4j.LoggerFactory;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.HttpURLConnection;
-import java.net.URI;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.Executors;
 import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
-import java.util.zip.GZIPOutputStream;
 
 import javax.annotation.Nullable;
 import javax.validation.constraints.NotNull;
-import javax.ws.rs.core.HttpHeaders;
-import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 /**
@@ -33,7 +21,7 @@ import javax.ws.rs.core.Response;
  *
  * @author Vikram Raman (vikram@wavefront.com)
  */
-public class WavefrontDirectSender implements WavefrontSender, Runnable {
+public class WavefrontDirectSender extends AbstractDirectConnectionHandler implements WavefrontSender {
 
   private static final String DEFAULT_SOURCE = "wavefrontDirectSender";
   private static final Logger LOGGER = LoggerFactory.getLogger(WavefrontDirectSender.class);
@@ -42,24 +30,10 @@ public class WavefrontDirectSender implements WavefrontSender, Runnable {
   private static final int MAX_QUEUE_SIZE = 50000;
   private static final int BATCH_SIZE = 10000;
 
-  private ScheduledExecutorService scheduler;
-  private final String server;
-  private final String token;
-  private DataIngesterAPI directService;
   private final LinkedBlockingQueue<String> buffer = new LinkedBlockingQueue<>(MAX_QUEUE_SIZE);
 
   public WavefrontDirectSender(String server, String token) {
-    this.server = server;
-    this.token = token;
-  }
-
-  @Override
-  public synchronized void connect() throws IllegalStateException, IOException {
-    if (directService == null) {
-      directService = new DataIngesterService(server, token);
-      scheduler = Executors.newScheduledThreadPool(1, new NamedThreadFactory(DEFAULT_SOURCE));
-      scheduler.scheduleAtFixedRate(this, 1, 1, TimeUnit.SECONDS);
-    }
+    super(server, token);
   }
 
   @Override
@@ -128,11 +102,7 @@ public class WavefrontDirectSender implements WavefrontSender, Runnable {
   }
 
   @Override
-  public void flush() throws IOException {
-    internalFlush();
-  }
-
-  private void internalFlush() throws IOException {
+  void internalFlush() throws IOException {
 
     if (!isConnected()) {
         return;
@@ -145,7 +115,7 @@ public class WavefrontDirectSender implements WavefrontSender, Runnable {
 
     Response response = null;
     try (InputStream is = pointsToStream(points)) {
-      response = directService.report("graphite_v2", is);
+      response = report("graphite_v2", is);
       if (response.getStatusInfo().getFamily() == Response.Status.Family.SERVER_ERROR ||
           response.getStatusInfo().getFamily() == Response.Status.Family.CLIENT_ERROR) {
         LOGGER.debug("Error reporting points, respStatus=" + response.getStatus());
@@ -184,26 +154,8 @@ public class WavefrontDirectSender implements WavefrontSender, Runnable {
   }
 
   @Override
-  public synchronized boolean isConnected() {
-    return directService != null;
-  }
-
-  @Override
   public int getFailureCount() {
     return 0;
-  }
-
-  @Override
-  public synchronized void close() throws IOException {
-    if (directService != null) {
-      try {
-        scheduler.shutdownNow();
-      } catch (SecurityException ex) {
-        LOGGER.debug("shutdown error", ex);
-      }
-      scheduler = null;
-      directService = null;
-    }
   }
 
   @Override
@@ -212,74 +164,6 @@ public class WavefrontDirectSender implements WavefrontSender, Runnable {
       this.internalFlush();
     } catch (Throwable ex) {
       LOGGER.debug("Unable to report to Wavefront", ex);
-    }
-  }
-
-  private static final class DataIngesterService implements DataIngesterAPI {
-
-    private final String token;
-    private final URI uri;
-    private static final String BAD_REQUEST = "Bad client request";
-    private static final int CONNECT_TIMEOUT = 30000;
-    private static final int READ_TIMEOUT = 10000;
-
-    public DataIngesterService(String server, String token) throws IOException  {
-      this.token = token;
-      uri = URI.create(server);
-    }
-
-    @Override
-    public Response report(String format, InputStream stream) throws IOException {
-
-      /**
-       * Refer https://docs.oracle.com/javase/8/docs/technotes/guides/net/http-keepalive.html
-       * for details around why this code is written as it is.
-       */
-
-      int statusCode = 400;
-      String respMsg = BAD_REQUEST;
-      HttpURLConnection urlConn = null;
-      try {
-        URL url = new URL(uri.getScheme(), uri.getHost(), uri.getPort(), String.format("/report?f=" + format));
-        urlConn = (HttpURLConnection) url.openConnection();
-        urlConn.setDoOutput(true);
-        urlConn.addRequestProperty(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_OCTET_STREAM);
-        urlConn.addRequestProperty(HttpHeaders.CONTENT_ENCODING, "gzip");
-        urlConn.addRequestProperty(HttpHeaders.AUTHORIZATION, "Bearer " + token);
-
-        urlConn.setConnectTimeout(CONNECT_TIMEOUT);
-        urlConn.setReadTimeout(READ_TIMEOUT);
-
-        try (GZIPOutputStream gzipOS = new GZIPOutputStream(urlConn.getOutputStream())) {
-          byte[] buffer = new byte[4096];
-          int len = 0;
-          while ((len = stream.read(buffer)) > 0) {
-            gzipOS.write(buffer);
-          }
-          gzipOS.flush();
-        }
-        statusCode = urlConn.getResponseCode();
-        respMsg = urlConn.getResponseMessage();
-        readAndClose(urlConn.getInputStream());
-      } catch (IOException ex) {
-        if (urlConn != null) {
-          statusCode = urlConn.getResponseCode();
-          respMsg = urlConn.getResponseMessage();
-          readAndClose(urlConn.getErrorStream());
-        }
-      }
-      return Response.status(statusCode).entity(respMsg).build();
-    }
-
-    private void readAndClose(InputStream stream) throws IOException {
-      if (stream != null) {
-        try (InputStream is = stream) {
-          byte[] buffer = new byte[4096];
-          int ret = 0;
-          // read entire stream before closing
-          while ((ret = is.read(buffer)) > 0) {}
-        }
-      }
     }
   }
 }
